refactor: break @confect/cli ↔ @confect/server workspace cycle
Move the `DatabaseSchema` brand tag, `Any` interface, and predicate into
`@confect/core/DatabaseSchema` and re-export them from `@confect/server`.
`@confect/cli` now imports the predicate from `@confect/core` and no
longer peer-depends on `@confect/server`, eliminating the cyclic
workspace dependency warning emitted by pnpm install.

Also renames `DatabaseSchema.isSchema` to `DatabaseSchema.isDatabaseSchema`
to match the `is<TypeName>` convention used elsewhere in the codebase.
The underlying brand string is unchanged, so existing schema values
continue to be recognized.
