+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
+export * as DatabaseSchema from "./DatabaseSchema";
 export * as FunctionProvenance from "./FunctionProvenance";
 export * as FunctionSpec from "./FunctionSpec";
 export * as GenericId from "./GenericId";
